app/backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
email-validator>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6

# LLM
litellm>=1.17.0

# Utilities
httpx>=0.26.0
python-dateutil>=2.8.2

# Image handling
pillow>=10.2.0
python-magic>=0.4.27
